% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_sf.geo_list}
\alias{as_sf.mapzen_vector_layer}
\title{Coerce a Mapzen response to a simple features object}
\usage{
as_sf(geo, ...)

\method{as_sf}{geo_list}(geo, ...)

\method{as_sf}{mapzen_vector_layer}(geo, ...)
}
\arguments{
\item{geo}{The object to be converted}

\item{...}{not currently used}
}
\description{
Coerces responses to class sf. See \code{vignette("sf1", package = "sf")} for
more information about Simple Features for R.
}
