% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-sp.R
\name{as_sp}
\alias{as_sp}
\title{Coerce a Mapzen response to an Spatial*DataFrame}
\usage{
as_sp(geo, ...)
}
\arguments{
\item{geo}{The object to be converted}

\item{...}{not currently used}

\item{geometry_type}{"point", "line", or "polygon" -- can be left NULL and only
needs to be specified when an object contains muliple geometry types.}
}
\description{
Coerces responses to SpatialPoints, SpatialLines, or SpatialPolygons data
frames.
}

