% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_tiles}
\alias{ca_tiles}
\title{Vector tiles the contain California}
\format{
An object of class \code{mapzen_vector_tiles} (inherits from \code{list}) of length 9.
}
\source{
Mapzen, OpenStreetMap contributors, Who's On First, Natural Earth,
and openstreetmapdata.com
}
\usage{
ca_tiles
}
\description{
Vector tiles the contain California
}
\keyword{datasets}
