% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-references.R
\docType{data}
\name{mapzen_references}
\alias{mapzen_references}
\alias{mz_sources}
\alias{mz_layers}
\alias{mz_countries}
\title{Reference lists}
\format{An object of class \code{list} of length 8.}
\usage{
mz_sources

mz_layers

mz_countries
}
\description{
Lists of sources, layers, and countries, as they are expected to appear in
the \code{\link{mz_search}} functions.
}
\keyword{datasets}
