% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marina_walks}
\alias{marina_walks}
\title{Pedestrian isochrones from the Berkeley Marina for 10 and 15 minutes}
\format{An object of class \code{mapzen_isochrone_list} (inherits from \code{geo_list}) of length 3.}
\source{
Mapzen, OpenStreetMap, British Oceanographic Data Centre,
NASA, USGS, and Transitland.
}
\usage{
marina_walks
}
\description{
Isochrone results from Mapzen as of January 8, 2017. The location for the
isochrones is the Berkeley Marina, lat 37.86613, lon -122.3151
}
\keyword{datasets}
