% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marina_walks_polygons}
\alias{marina_walks_polygons}
\title{Pedestrian isochrones from the Berkeley Marina for 10 and 15 minutes, as polygons}
\format{
An object of class \code{mapzen_isochrone_list} (inherits from \code{geo_list}) of length 3.
}
\source{
Mapzen, OpenStreetMap, British Oceanographic Data Centre,
NASA, USGS, and Transitland.
}
\usage{
marina_walks_polygons
}
\description{
Polygon Isochrone results (using \code{polygons = TRUE}) from Mapzen as of
January 10, 2017. The location for the isochrones is the Berkeley Marina,
lat 37.86613, lon -122.3151, and the contours are 10 and 15 minutes for a
pedestrian costing model.
}
\keyword{datasets}
