% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-accessors.R
\name{mz_attribution}
\alias{mz_attribution}
\title{Get attribution information from a \code{mapzen_geo_list}}
\usage{
mz_attribution(geo)
}
\arguments{
\item{geo}{A mapzen geo list}
}
\description{
Get attribution information from a \code{mapzen_geo_list}
}

