% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-bbox.R
\name{mz_bbox}
\alias{mz_bbox}
\alias{mz_bbox.mapzen_geo_list}
\alias{mz_bbox.mazpen_isochrone_list}
\title{Get the bounding box}
\usage{
mz_bbox(geo)

\method{mz_bbox}{mapzen_geo_list}(geo)

\method{mz_bbox}{mazpen_isochrone_list}(geo)
}
\arguments{
\item{geo}{A mapzen geo list or isochrone list}
}
\value{
A tibble
}
\description{
Returns the bottom left and top right corners of the box that contains a
mapzen object (mz_geo_list or mz_isochrone_list). The returned value can be
used directly as the \code{boundary.rect} parameter for
\code{\link{search}} functions.
}

