% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage-statistics.R
\name{mz_check_usage}
\alias{mz_check_usage}
\title{Check usage statistics}
\usage{
mz_check_usage()
}
\description{
Prints out remaining queries for various time periods. \code{rmapzen} manages
rate limiting for the per-second limits, but does not keep track of the daily
limits.
}
\details{
This function is populated from the headers of responses to various API requests.
If no queries have been made, or if the only queries so far have hit cache servers,
then no information will be available.
}
