% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-helpers.R
\name{mz_contours}
\alias{mz_contours}
\title{Create an mz_contours object}
\usage{
mz_contours(times, colors = NULL)
}
\arguments{
\item{times}{Times in minutes for the contour. Up to a maximum of 4 numbers.}

\item{colors}{Colors for the contours. By default, a palette will be constructed
from the Colorbrewer 4-class oranges palette.}
}
\description{
Contours are given as inputs to \code{\link{mz_isochrone}}. This function
makes it convenient to construct them.
}
