% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-accessors.R
\name{mz_coordinates}
\alias{mz_coordinates}
\title{Extract a data frame of coordinates from a \code{mapzen_geo_list}}
\usage{
mz_coordinates(geo)
}
\arguments{
\item{geo}{A mapzen geo list}
}
\value{
A tibble, with columns \code{lon} and \code{lat}
}
\description{
Extract a data frame of coordinates from a \code{mapzen_geo_list}
}

