% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-helpers.R
\name{mz_date_time}
\alias{mz_date_time}
\title{Create mz_date_time objects}
\usage{
mz_date_time(date_time, type = "departure")
}
\arguments{
\item{date_time}{A POSIXt date-time object}

\item{type}{"departure" or "arrival"}
}
\description{
Mobility services (such as \code{mz_isochrone}) take, optionally, a date_time
argument that specifies the date and time along with type (departure/arrival).
This function constructs the appropriate objects to use as date_time arguments.
}
