% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{mz_isochrone}
\alias{mz_isochrone}
\title{Retrieve isochrones}
\usage{
mz_isochrone(locations, costing_model, contours, date_time = NULL,
  polygon = NULL, denoise = NULL, generalize = NULL, id = "my-iso",
  api_key = mz_key())
}
\description{
Retrieve isochrones
}

