% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-helpers.R
\name{mz_location}
\alias{mz_location}
\alias{as.mz_location}
\alias{as.mz_location.default}
\alias{as.mz_location.mz_geocode_result}
\title{Create/extract lat/lon location information}
\usage{
mz_location(lat, lon)

as.mz_location(x, ...)

\method{as.mz_location}{default}(x, ...)

\method{as.mz_location}{mz_geocode_result}(x, ...)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{x}{An object that has location information}

\item{...}{Not currently used}
}
\description{
\code{mz_location} constructs a new \code{mz_location} object, which can be
used with functions such as \code{\link{mz_isochrone}} or
\code{\link{mz_reverse_geocode}}. \code{as.mz_location} coerces eligible
objects to \code{mz_location}s.
}
\seealso{
\code{\link{mz_isochrone}} For using the Mapzen isochrone service
\code{\link{mz_contours}}, \code{\link{mz_costing}}, and
\code{\link{mz_costing_options}} for other argument constructors
}
