% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-normalize-addr.R
\name{mz_normalize_addr}
\alias{mz_normalize_addr}
\alias{mz_normalise_addr}
\title{Normalize Address with Mapzen Libpostal API}
\usage{
mz_normalize_addr(address, api_key = mz_key())

mz_normalise_addr(address, api_key = mz_key())
}
\arguments{
\item{address}{Address to normalize}

\item{api_key}{Your Mapzen API key. The default is to look for the value in
the MAPZEN_KEY environment variable.}
}
\description{
Normalize Address with Mapzen Libpostal API
}
\examples{
\dontrun{
mz_normalize_addr("Quatre-vingt-douze Ave des Champs-\\u00c9lys\\u00e9es")
}

}
