% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-parse-addr.R
\name{mz_parse_addr}
\alias{mz_parse_addr}
\title{Parse Address with Mapzen Libpostal API}
\usage{
mz_parse_addr(address, format_keys = FALSE, api_key = mz_key())
}
\arguments{
\item{address}{Address to parse.}

\item{format_keys}{By default the Libpostal API returns results as a list of dictionaries, each containing a label and value key. This is because there are occasions when a given key may have multiple values, for example an address that contains a cross-street. If you would prefer to have API results returned as a simple dictionary with labels as keys and values as lists of possible strings you should set the \code{format_keys} parameter to \code{TRUE}.}

\item{api_key}{Your Mapzen API key. The default is to look for the value in
the \code{MAPZEN_KEY} environment variable.}
}
\description{
Parse Address with Mapzen Libpostal API
}
\examples{
\dontrun{
mz_parse_addr("The Book Club 100-106 Leonard St Shoreditch London EC2A 4RH, United Kingdom")
}
}
