% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{mz_provider}
\alias{mz_provider}
\title{Configure provider information}
\usage{
mz_provider(hostname, path = NULL, key = NULL, scheme = "https")
}
\arguments{
\item{hostname}{The hostname in the API URL, for instance \code{www.example.com}}

\item{path}{Specific path that all API requests must include, e.g. "v1"}

\item{key}{API key for this provider, if required}

\item{scheme}{The scheme for the URL, should always be "https"}
}
\description{
rmapzen works with most implementations of PELIAS. This function defines the
base URL for a particular API provider, and can be used to provider the
\code{provider} argument to \code{\link{mz_set_host}}.
}
\seealso{
\code{\link{mz_set_host}}
}
