% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-tile-coordinates.R
\name{mz_tile_coordinates}
\alias{as.mz_tile_coordinates}
\alias{as.mz_tile_coordinates.mz_bbox}
\alias{as.mz_tile_coordinates.mz_tile_coordinates}
\alias{mz_tile_coordinates}
\title{Specify tile coordinates}
\usage{
mz_tile_coordinates(x, y, z)

as.mz_tile_coordinates(obj, ...)

\method{as.mz_tile_coordinates}{mz_tile_coordinates}(obj, ...)

\method{as.mz_tile_coordinates}{mz_bbox}(obj, ..., height = 375,
  width = 500)
}
\arguments{
\item{x}{integer vector of x-coordinates}

\item{y}{integer vector of y-coordinates}

\item{z}{integer between 0 and 19 specifying the zoom level}

\item{obj}{An object that can be converted to tile coordinates}

\item{...}{Other arguments passed on to methods}

\item{height}{Height in pixels}

\item{width}{Width in pixels}
}
\description{
Specify tile coordinates
}

