% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-tiles.R
\name{mz_vector_tiles}
\alias{mz_vector_tiles}
\title{Request vector tile data}
\usage{
mz_vector_tiles(tile_coordinates, ...)
}
\arguments{
\item{tile_coordinates}{an \code{\link{mz_tile_coordinates}} object, or something
that can be coerced to one (including the output of \code{\link{mz_bbox}})}

\item{...}{Arguments passed on to \code{\link{as.mz_tile_coordinates}}

Multiple tiles will be stitched together and returned as one object.}
}
\description{
Request vector tile data
}

