% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oakland_public}
\alias{oakland_public}
\title{25 search results for "Oakland Public library branch"}
\format{
A \code{mapzen_geo_list} with 25 locations
}
\source{
Mapzen, OpenStreetMap, OpenAddresses, GeoNames, WhosOnFirst,
see \url{https://www.mapzen.com/rights/}
}
\usage{
oakland_public
}
\description{
Contains the search results from Mapzen's search service for the query
"Oakland public library branch" as of January 8, 2017.
}
\keyword{datasets}
