% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapzen-package.R
\docType{package}
\name{rmapzen-package}
\alias{rmapzen}
\alias{rmapzen-package}
\title{rmapzen: Client for 'Mapzen' and Related Map APIs}
\description{
Provides an interface to 'Mapzen'-based APIs (including geocode.earth, Nextzen, and NYC GeoSearch) for geographic search and geocoding, isochrone calculation, and vector data to draw map tiles. See \url{https://www.mapzen.com/documentation/} for more information. The original Mapzen has gone out of business, but 'rmapzen' can be set up to work with any provider who implements the Mapzen API.
}
\section{Search}{

All functionality described in \url{https://github.com/pelias/documentation/}
are supported:

\itemize{
    \item \code{\link{mz_search}}
    \item \code{\link{mz_reverse_geocode}}
    \item \code{\link{mz_autocomplete}}
    \item \code{\link{mz_place}}
    \item \code{\link{mz_structured_search}}
}

Additionally, \code{\link{mz_geocode}} is useful for a common application of
search, that of just obtaining latitude and longitude for a given address or place.
}

\section{Isochrone}{

Isochrones are the areas reachable from a given location within a specified
period of time. Mapzen's Isochrone service can calculate isochrones for driving,
walking, cycling, or multimodal forms of transport:

\itemize{
    \item \code{\link{mz_isochrone}}
    \item \code{\link{mz_costing}}: for constructing "costing models" that describe
    method of transport along with speed and other options relevant to the
    calculation of the isochrone
    \item \code{\link{mz_costing_options}}: for selecting specific options when
    constructing a costing model
}
}

\section{Vector Tiles}{

\itemize{
    \item \code{\link{mz_vector_tiles}}: Request one or more adjacent tiles.
    Multiple map tiles will be stitched together before being returned as a single
    object.
    \item \code{\link{mz_tile_coordinates}}: When using \code{\link{mz_vector_tiles}},
    you must specify the geographic area for which you want tile data. One way to do so
    is using the x, y, z tile naming system (see \url{https://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}).
    \item \code{\link{mz_rect}}: Alternatively, you can specify the lower left
    and top-right points of a bounding box, which will automatically be converted
    to tile-coordinates when you use \code{\link{mz_vector_tiles}}
    \item \code{\link{mz_bbox}}: This is a generic function which will return
    the bounding box of any Mapzen object. In this way, you can request vector
    tiles for a region defined as the bounding box of an existing object.
}
}

\section{Data types and conversion}{

Objects returned by rmapzen can be converted to
simple features (sf) via the generic function
 \code{\link{as_sf}}. Search
and Isochrone objects can additionally be converted to ordinary data.frames
via \code{as.data.frame}.
}

\seealso{
\itemize{
    \item \url{https://tarakc02.github.io/rmapzen/} contains detailed examples
    \item \url{https://www.mapzen.com/documentation/} 'Mapzen' documentation
}
}
\author{
\strong{Maintainer}: Tarak Shah \email{tarak.shah@gmail.com}

Other contributors:
\itemize{
  \item Daniel Possenriede [contributor]
}

}
\keyword{'Mapzen'}
\keyword{'Mapzen',}
\keyword{(\url{https://github.com/pelias/documentation/}),}
\keyword{(\url{https://tilezen.readthedocs.io/en/latest/})}
\keyword{(\url{https://valhalla.readthedocs.io/en/latest/}),}
\keyword{A}
\keyword{API.}
\keyword{Isochrone}
\keyword{Search}
\keyword{The}
\keyword{Tile}
\keyword{Vector}
\keyword{and}
\keyword{application}
\keyword{client}
\keyword{following}
\keyword{for}
\keyword{from}
\keyword{functions:}
\keyword{interfaces}
\keyword{internal}
\keyword{package}
\keyword{provides}
\keyword{rmapzen}
\keyword{rmapzen:}
\keyword{services}
\keyword{the}
\keyword{to}
\keyword{via}
