% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-autocomplete.R, R/mz-place.R, R/mz-reverse.R, R/mz-search.R
\name{mz_autocomplete}
\alias{mz_autocomplete}
\alias{mz_place}
\alias{mz_reverse_geocode}
\alias{mz_search}
\alias{search}
\title{Mapzen search API}
\usage{
mz_autocomplete(text, boundary.country = NULL, boundary.rect = NULL,
  focus.point = NULL, sources = NULL, layers = NULL, api_key = mz_key())

mz_place(ids, ..., api_key)

mz_reverse_geocode(point, size = NULL, layers = NULL, sources = NULL,
  boundary.country = NULL, api_key = mz_key())

mz_search(text, size = 10, boundary.country = NULL, boundary.rect = NULL,
  boundary.circle = NULL, focus.point = NULL, sources = NULL,
  layers = NULL, api_key = mz_key())
}
\description{
Functions to access the various endpoints from the Mapzen Search API.
For more details, see \url{https://mapzen.com/documentation/search/}.
}

