% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-sp.R
\name{as_sp}
\alias{as_sp}
\title{Coerce a Mapzen response to an Spatial*DataFrame}
\usage{
as_sp(geo, ...)
}
\arguments{
\item{geo}{The object to be converted}

\item{...}{not currently used}
}
\description{
Coerces responses to SpatialPoints (for search responses) or SpatialLines
(for isochrone responses) data frames.
}

