% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz-coordinates.R
\name{mz_coordinates}
\alias{mz_coordinates}
\alias{mz_coordinates.mapzen_geo_list}
\title{Extract a data frame of coordinates from a \code{mapzen_geo_list}}
\usage{
mz_coordinates(geo)

\method{mz_coordinates}{mapzen_geo_list}(geo)
}
\arguments{
\item{geo}{A mapzen geo list}
}
\value{
A tibble, with columns \code{lon} and \code{lat}.
}
\description{
Extract a data frame of coordinates from a \code{mapzen_geo_list}
}
\examples{
mz_coordinates(oakland_public)

}
